// server/src/models/User.ts
/** @format */

export type UserType = {
  id?: number; // Optional: as it's auto-generated by PostgreSQL
  name: string;
  email: string;
  password: string; // Remember to handle password hashing appropriately
  createdAt?: Date; // Optional: as it's auto-generated or set by trigger
  updatedAt?: Date; // Optional: as it's auto-generated or set by trigger
  phone?: string;
  address?: string[];
  city?: string;
  state?: string;
  zip?: string;
  country?: string;
};